refactor: 코드 useHooks 구조로 교체, title Object에 맞게 코드 수정
