+++ src/components/MainSection/Like/style.js
