+++ src/components/MainSection/PostDetail/Comments/index.jsx
@@ -6,34 +6,10 @@ import { createComment } from '../../../../apis/comment';
 import { getChannelPosts, getPostDetail } from '../../../../apis/post';
 import { CommentCount, InputContainer, CommentContainer } from './style';
 import { createAlarm } from '../../../../apis/alarm';
+import useComment from '../../../../hooks/useComment';
 
 const Comments = ({ postId, comments }) => {
-  const [postDetail, setPostDetail] = useRecoilState(postStateFamily(postId));
-  const [comment, setComment] = useState('');
-
-  const onChange = (e) => {
-    setComment(e.target.value);
-  };
-
-  const onSubmit = async (e) => {
-    e.preventDefault();
-
-    if (comment.length > 1) {
-      const result = await createComment(postId, comment);
-
-      if (result.status === 200) {
-        setComment('');
-
-        //알림보내기
-        const { _id, post } = result.data;
-        await createAlarm('COMMENT', _id, postDetail.post.author._id, post);
-
-        //PostDetail atom 업데이트
-        const { data } = await getPostDetail(postId);
-        setPostDetail({ key: postId, post: data });
-      }
-    }
-  };
+  const [comment, onChange, onSubmit] = useComment(postId);
 
   return (
     <>
