+++ src/components/MainSection/PostDetail/index.jsx
@@ -1,55 +1,29 @@
-import { useEffect } from 'react';
-import { useRecoilValue, useRecoilState } from 'recoil';
-import { selectedPostState, postStateFamily } from '../../../recoil/postStates';
-import { getPostDetail } from '../../../apis/post';
 import AuthorInfo from '../Post/AuthorInfo';
 import LeftImage from '../Post/LeftImage';
 import Tags from '../Post/Tags';
-import Heart from '../Heart';
+import Heart from '../Like';
 import Comments from './Comments';
 import Skeleton from '../../common/Skeleton';
+import usePostDetail from '../../../hooks/usePostDetail';
 import {
   PostDetailContainer,
   RightContainer,
   RightContainerContent,
   Title,
   BottomContainer,
+  Country,
+  Content,
 } from './style';
 
 const PostDetail = () => {
-  const postId = useRecoilValue(selectedPostState);
-  const [{ post, content, titleData }, setPostDetail] = useRecoilState(postStateFamily(postId));
-
-  const getPostData = async () => {
-    if (post === null) {
-      const { data } = await getPostDetail(postId);
-      const { title } = data;
-      let content;
-      let titleData;
-
-      if (title.includes('{')) {
-        content = JSON.parse(title).content;
-        titleData = JSON.parse(title).title;
-      } else {
-        titleData = title.split('/')[0];
-      }
-
-      setPostDetail({ key: postId, post: data, content, titleData });
-    }
-  };
-
-  useEffect(() => {
-    if (postId) {
-      getPostData();
-    }
-  }, [postId]);
+  const { post } = usePostDetail();
 
   return post ? (
     <PostDetailContainer>
       <LeftImage src={post.image} width={'50%'} height={'100%'} style={{ borderRadius: '16px' }} />
       <RightContainer>
         <RightContainerContent alignItem="flex-start" style={{ paddingBottom: '60px' }}>
-          <Title>{titleData}</Title>
+          <Title>{post.title}</Title>
         </RightContainerContent>
 
         <RightContainerContent
@@ -59,22 +33,25 @@ const PostDetail = () => {
           style={{ paddingBottom: '20px' }}
         >
           <AuthorInfo image={post.author.image} fullName={post.author.fullName} />
-          <div>벨기에, 브리쉘</div>
+          <Country>
+            <img src={require('../../../../static/images/Location.png')} />
+            <span>{post.country}</span>
+          </Country>
         </RightContainerContent>
         <RightContainerContent flexDirection="row" justifyContent="space-between">
-          <Tags data={post.title} alignItem="flex-end" />
+          <Tags
+            date={post.date}
+            gender={post.gender}
+            personnel={post.personnel}
+            alignItem="flex-end"
+          />
           <RightContainerContent flexDirection="row">
-            <Heart
-              likes={post.likes}
-              author={post.author}
-              postId={post._id}
-              styleProps={{ fontSize: '25px' }}
-            />
+            <Heart likes={post.likes} author={post.author} postId={post._id} />
           </RightContainerContent>
         </RightContainerContent>
-        <div>{content}</div>
+        <Content>{post.content}</Content>
         <BottomContainer alignItem="flex-start">
-          <Comments postId={postId} comments={post.comments} />
+          <Comments postId={post._id} comments={post.comments} />
         </BottomContainer>
       </RightContainer>
     </PostDetailContainer>
