+++ src/utils/sleep.js
@@ -1,0 +1,3 @@
+export const sleep = (ms) => {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+};
